frontend/requirements.txt
Django==5.2.6
djangorestframework>=3.15.0,<4.0

requests==2.32.3
django-allauth==0.60.1


gunicorn==21.2.0
psycopg2-binary==2.9.9  # If using PostgreSQL

python-decouple==3.8  # Add this line